se agrega la clase pais
se agrega pais para poder relacionar en que parte del pais se encuentra la universidad
