+++ src/main/java/universidades/Nelson/Class/City.java
@@ -1,0 +1,27 @@
+package universidades.Nelson.Class;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "city")
+public class City {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_city",unique = true,nullable = false)
+    private Long id ;
+    private String nombre_del_pais;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "id_country")
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    private Countries country;
+
+    public Countries getCountry() {
+        return country;
+    }
+
+    public void setCountry(Countries country) {
+        this.country = country;
+    }
+}
